fix: colocate purge logic
